+++ changelog.md
@@ -7,3 +7,6 @@
 
 * 2.5.0
    - Add an Erlang implementation of the Avro CRC 64 fingerprint algorithm, avro:crc64_fingerprint/1.
+   - Add `avro:make_simple_decoder/2` and `avro:make_simple_encoder/2` to simplify most common use cases.
+     `avro:make_decoder/2` and `avro:make_encoder/2` are not deprecated and can still be used when
+     there is a need to use the same decoer/encoder for multiple schema.